requirements.txt
pyserial==3.5
websocket-client==1.4.2
requests~=2.28.2
PySocks==1.7.1
